ch-22.3.58346.58453

+++ BaseApp/Source/Base Application/CRM/ArchiveManagement.Codeunit.al
@@ -252,7 +252,7 @@ codeunit 5063 ArchiveManagement
         RestoreDocument: Boolean;
         OldOpportunityNo: Code[20];
         IsHandled: Boolean;
-        DoCheck: Boolean;
+        DoCheck, SkipDeletingLinks : Boolean;
     begin
         OnBeforeRestoreSalesDocument(SalesHeaderArchive, IsHandled);
         if IsHandled then
@@ -315,8 +315,10 @@ codeunit 5063 ArchiveManagement
                 OldOpportunityNo := SalesHeader."Opportunity No.";
                 SalesHeader."Opportunity No." := '';
             end;
-            OnRestoreDocumentOnBeforeDeleteSalesHeader(SalesHeader);
-            SalesHeader.DeleteLinks();
+            SkipDeletingLinks := false;
+            OnRestoreDocumentOnBeforeDeleteSalesHeader(SalesHeader, SkipDeletingLinks);
+            if not SkipDeletingLinks then
+                SalesHeader.DeleteLinks();
             SalesHeader.Delete(true);
             OnRestoreDocumentOnAfterDeleteSalesHeader(SalesHeader);
 
@@ -741,6 +743,7 @@ codeunit 5063 ArchiveManagement
         SalesCommentLine."Line No." := NextLine;
         SalesCommentLine.Date := WorkDate();
         SalesCommentLine.Comment := StrSubstNo(Text004, Format(SalesHeaderArchive."Version No."));
+        OnRestoreSalesLineCommentsOnBeforeInsertSalesCommentLine(SalesCommentLine);
         SalesCommentLine.Insert();
     end;
 
@@ -955,7 +958,7 @@ codeunit 5063 ArchiveManagement
     end;
 
     [IntegrationEvent(false, false)]
-    local procedure OnRestoreDocumentOnBeforeDeleteSalesHeader(var SalesHeader: Record "Sales Header")
+    local procedure OnRestoreDocumentOnBeforeDeleteSalesHeader(var SalesHeader: Record "Sales Header"; var SkipDeletingLinks: Boolean)
     begin
     end;
 
@@ -1021,6 +1024,11 @@ codeunit 5063 ArchiveManagement
 
     [IntegrationEvent(false, false)]
     local procedure OnBeforeArchSalesDocumentNoConfirm(var SalesHeader: Record "Sales Header"; var IsHandled: Boolean)
+    begin
+    end;
+
+    [IntegrationEvent(false, false)]
+    local procedure OnRestoreSalesLineCommentsOnBeforeInsertSalesCommentLine(var SalesCommentLine: Record "Sales Comment Line")
     begin
     end;
 }
